AG-16328 Tidy up: move glob-util function to api-codegen.spec.ts
