+++ packages/ag-charts-website/e2e/api-codegen.spec.ts
@@ -1,7 +1,25 @@
+import fs from 'fs';
+import glob from 'glob';
 import path from 'path';
 
 import { expect, test } from './fixture';
-import { createConsoleLogs, getPagesWithAPIReferences, gotoUrl, setupIntrinsicAssertions } from './util';
+import { createConsoleLogs, gotoUrl, setupIntrinsicAssertions, toPageUrl } from './util';
+
+function findPagesWithAPIReferences(): string[] {
+    // Regex that matches `{% apiReference ... /%}` even across multiple lines
+    const apiRefRegex = /\{%\s*apiReference[\s\S]*?\/%\}/m;
+
+    const pathsWithApiRef = glob.sync('./src/content/docs/**/index.mdoc').filter((filepath) => {
+        const content = fs.readFileSync(filepath, 'utf8');
+        return apiRefRegex.test(content);
+    });
+
+    // Extract the `**` part of the glob.
+    return pathsWithApiRef.map((filepath) => {
+        const dir = path.dirname(filepath);
+        return path.relative('./src/content/docs', dir);
+    });
+}
 
 test.use({ viewport: { width: 1400, height: 900 } });
 
@@ -16,13 +34,10 @@ test.describe('api-codegen', () => {
             expect(buttonCount).toBeGreaterThan(0);
         });
 
-        const urls: string[] = getPagesWithAPIReferences();
-        console.log(urls);
-
-        for (const url of urls) {
-            const basename = path.basename(url);
-            test(basename, async ({ page }) => {
-                await gotoUrl(page, url);
+        const pageNames: string[] = findPagesWithAPIReferences();
+        for (const name of pageNames) {
+            test(name, async ({ page }) => {
+                await gotoUrl(page, toPageUrl(`javascript/${name}`));
 
                 const buttons = page.locator('button[class*="_seeMore"]');
                 const count = await buttons.count();
